requirements.txt
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
streamlit